replaced old DB with new DB and hooked it up to the server
